dice que desconoce la DB

+++ src/database/models/Product-color.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, dataTypes) => {
     let alias = 'ProductColor';
     let cols = {
         id: {
-            type: dataTypes.INT(10).UNSIGNED,
+            type: dataTypes.BIGINT(10).UNSIGNED,
             primaryKey: true,
             autoIncrement: true
         },
