+++ src/database/models/Product-stock.js
@@ -2,22 +2,22 @@ module.exports = (sequelize, dataTypes) => {
     let alias = 'ProductStock';
     let cols = {
         id: {
-            type: dataTypes.INT(10).UNSIGNED,
+            type: dataTypes.BIGINT(10).UNSIGNED,
             primaryKey: true,
             autoIncrement: true
         },        
        
         id_product: {
-            type: dataTypes.INT(10),
+            type: dataTypes.BIGINT(10),
             allowNull: true
         },
 
         quantity: {
-            type: dataTypes.INT(8),
+            type: dataTypes.INTEGER,
             allowNull: true
         },  
         break_point: {
-            type: dataTypes.INT(8),
+            type: dataTypes.INTEGER,
             allowNull: true
         }  
     };
